webapp/backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
pyyaml
